Minor changes
Backing up just to be on the safe side :)

+++ Animals.h
@@ -8,7 +8,7 @@
 class AnimalState
 {
 public:
-	int xPos, yPos;
+	float xPos, yPos;
 	int Frame, State;
 	int Height, Width;
 	int Radius;
